+++ Documentation/SwiftUI VIPER Architecture/Example Project/SwiftUIVIPERArchitectureDemo/Gateways/Posts/PostsNetworkGateway.swift
@@ -11,14 +11,22 @@ import VCore
 // MARK: - Posts Network Gateway
 struct PostsNetworkGateway: PostsGateway {
     func fetch() async throws -> PostsEntity {
-        var request: NetworkRequest = .init(url: "https://jsonplaceholder.typicode.com/posts")
-        request.method = .GET
-        
+        let urlString: String = "https://jsonplaceholder.typicode.com/posts"
+        guard let url: URL = .init(string: urlString) else { throw URLError(.badURL) }
+
+        var request: URLRequest = .init(url: url)
+        request.httpMethod = "GET"
+        try request.addHTTPHeaderFields(object: JSONRequestHeaders())
+
         try? await Task.sleep(seconds: 1)
-        
-        let posts: [PostsEntity.Post] = try await NetworkClient.default.object(from: request)
+
+        let (data, response): (Data, URLResponse) = try await URLSession.shared.data(for: request)
+
+        guard response.isSuccessHTTPStatusCode else { throw URLError(.badServerResponse) }
+
+        let posts: [PostsEntity.Post] = try JSONDecoder().decode([PostsEntity.Post].self, from: data)
         let entity: PostsEntity = .init(posts: posts)
-        
+
         return entity
     }
 }