+++ Sources/VCore/Models/HTTP Headers/JSONAuthorizedRequestHeaders.swift
@@ -10,10 +10,10 @@ import Foundation
 // MARK: - JSON Authorized Request Headers
 /// Authorized JSON request headers that pass `application/json` as `accept`, `application/json` as `contentType`, and `authorization` as `Bearer \(token)`.
 ///
-/// Can be used in `NetworkClient`.
+/// Can be used in `URLRequest`.
 ///
-///     var request: NetworkRequest = .init(url: ...)
-///     try request.addHeaders(object: JSONAuthorizedRequestHeaders(
+///     var request: URLRequest = ...
+///     try request.addHTTPHeaderFields(object: JSONAuthorizedRequestHeaders(
 ///         token: "token"
 ///     ))
 ///