+++ Tests/VCoreTests/Helpers/XCTest.XCTAssertEqualColor.swift
@@ -13,71 +13,65 @@ import VCore
 
 #if canImport(UIKit)
 
-extension XCTest {
-    func XCTAssertEqualColor(
-        _ expression1: @autoclosure () throws -> UIColor,
-        _ expression2: @autoclosure () throws -> UIColor,
-        _ message: @autoclosure () -> String = "",
-        file: StaticString = #filePath,
-        line: UInt = #line
-    ) {
-        let lhs = try! expression1().rgbaValues // Force-unwrap
-        let rhs = try! expression2().rgbaValues // Force-unwrap
-        
-        XCTAssertEqual(lhs.red, rhs.red, message(), file: file, line: line)
-        XCTAssertEqual(lhs.green, rhs.green, message(), file: file, line: line)
-        XCTAssertEqual(lhs.blue, rhs.blue, message(), file: file, line: line)
-        XCTAssertEqual(lhs.alpha, rhs.alpha, message(), file: file, line: line)
-    }
+func XCTAssertEqualColor(
+    _ expression1: @autoclosure () throws -> UIColor,
+    _ expression2: @autoclosure () throws -> UIColor,
+    _ message: @autoclosure () -> String = "",
+    file: StaticString = #filePath,
+    line: UInt = #line
+) {
+    let lhs = try! expression1().rgbaValues // Force-unwrap
+    let rhs = try! expression2().rgbaValues // Force-unwrap
+
+    XCTAssertEqual(lhs.red, rhs.red, message(), file: file, line: line)
+    XCTAssertEqual(lhs.green, rhs.green, message(), file: file, line: line)
+    XCTAssertEqual(lhs.blue, rhs.blue, message(), file: file, line: line)
+    XCTAssertEqual(lhs.alpha, rhs.alpha, message(), file: file, line: line)
 }
 
 #elseif canImport(AppKit)
 
-extension XCTest {
-    func XCTAssertEqualColor(
-        _ expression1: @autoclosure () throws -> NSColor,
-        _ expression2: @autoclosure () throws -> NSColor,
-        _ message: @autoclosure () -> String = "",
-        file: StaticString = #filePath,
-        line: UInt = #line
-    ) {
-        let lhs = try! expression1().rgbaValues // Force-unwrap
-        let rhs = try! expression2().rgbaValues // Force-unwrap
-        
-        XCTAssertEqual(lhs.red, rhs.red, message(), file: file, line: line)
-        XCTAssertEqual(lhs.green, rhs.green, message(), file: file, line: line)
-        XCTAssertEqual(lhs.blue, rhs.blue, message(), file: file, line: line)
-        XCTAssertEqual(lhs.alpha, rhs.alpha, message(), file: file, line: line)
-    }
+func XCTAssertEqualColor(
+    _ expression1: @autoclosure () throws -> NSColor,
+    _ expression2: @autoclosure () throws -> NSColor,
+    _ message: @autoclosure () -> String = "",
+    file: StaticString = #filePath,
+    line: UInt = #line
+) {
+    let lhs = try! expression1().rgbaValues // Force-unwrap
+    let rhs = try! expression2().rgbaValues // Force-unwrap
+
+    XCTAssertEqual(lhs.red, rhs.red, message(), file: file, line: line)
+    XCTAssertEqual(lhs.green, rhs.green, message(), file: file, line: line)
+    XCTAssertEqual(lhs.blue, rhs.blue, message(), file: file, line: line)
+    XCTAssertEqual(lhs.alpha, rhs.alpha, message(), file: file, line: line)
 }
 
 #endif
 
-extension XCTest {
-    @available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, *)
-    func XCTAssertEqualColor(
-        _ expression1: @autoclosure () throws -> Color,
-        _ expression2: @autoclosure () throws -> Color,
-        _ message: @autoclosure () -> String = "",
-        file: StaticString = #filePath,
-        line: UInt = #line
-    ) {
+@available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, *)
+func XCTAssertEqualColor(
+    _ expression1: @autoclosure () throws -> Color,
+    _ expression2: @autoclosure () throws -> Color,
+    _ message: @autoclosure () -> String = "",
+    file: StaticString = #filePath,
+    line: UInt = #line
+) {
 #if canImport(UIKit)
-        return XCTAssertEqualColor(
-            UIColor(try! expression1()), // Force-unwrap
-            UIColor(try! expression2()), // Force-unwrap
-            message(),
-            file: file,
-            line: line
-        )
+    return XCTAssertEqualColor(
+        UIColor(try! expression1()), // Force-unwrap
+        UIColor(try! expression2()), // Force-unwrap
+        message(),
+        file: file,
+        line: line
+    )
 #elseif canImport(AppKit)
-        return XCTAssertEqualColor(
-            NSColor(try! expression1()), // Force-unwrap
-            NSColor(try! expression2()), // Force-unwrap
-            message(),
-            file: file,
-            line: line
-        )
+    return XCTAssertEqualColor(
+        NSColor(try! expression1()), // Force-unwrap
+        NSColor(try! expression2()), // Force-unwrap
+        message(),
+        file: file,
+        line: line
+    )
 #endif
-    }
 }