+++ Tests/VCoreTests/Tests/Extensions/Foundation/Date/DateAgeTests.swift
@@ -13,7 +13,7 @@ final class DateAgeTests: XCTestCase {
     func testInvalidDate() {
         let calendar: Calendar = .current
 
-        guard let birthDate: Date = Calendar.current.date(byAdding: .second, value: 1, to: Date()) else { fatalError() }
+        guard let birthDate: Date = Calendar.current.date(byAdding: .second, value: 1, to: Date()) else { XCTFail(); return }
 
         XCTAssertNil(birthDate.age(inCalendar: calendar))
     }
@@ -21,7 +21,7 @@ final class DateAgeTests: XCTestCase {
     func testValidDate() {
         let calendar: Calendar = .current
         
-        guard let birthDate: Date = calendar.date(from: DateComponents(year: 1970, month: 1, day: 1)) else { fatalError() }
+        guard let birthDate: Date = calendar.date(from: DateComponents(year: 1970, month: 1, day: 1)) else { XCTFail(); return }
         
         XCTAssertNotNil(birthDate.age(inCalendar: calendar))
     }